frontend/requirements.txt
Flask==1.0.2
google-cloud-logging==1.10.0
google-cloud-storage==1.13.2